Implement food api data source search functionality
- Return a list of food entities given the results of an API query to
  the FoodData Central API
- Modify food entity serving and nutrition fields from int to double
- Implement FoodApiDataSource as a seperate datasource

+++ lib/features/food/domain/entities/food_entity.dart
@@ -4,12 +4,12 @@ import 'package:equatable/equatable.dart';
 /// ie. banana, dumpling, chicken pot pie
 class FoodEntity extends Equatable {
   final String name;
-  final int servingSize;
+  final double servingSize;
   final String servingSizeUnit;
-  final int calories;
-  final int carbs;
-  final int protein;
-  final int fat;
+  final double calories;
+  final double carbs;
+  final double protein;
+  final double fat;
 
   const FoodEntity(
       {required this.name,
